Made Table.js reusable; added conversion logic to App.js
