+++ src/App.js
@@ -1,6 +1,24 @@
 import React, { Component } from "react";
 import "./App.css";
-import Table from "./Table";
+import Table from "./components/Table";
+import data, { getAirlineById, getAirportByCode } from "./data.js";
+const { routes, airlines, airports } = data;
+
+// Columns that correspond with route objects
+const columns = [
+  { name: "Airline", property: "airline" },
+  { name: "Source Airport", property: "src" },
+  { name: "Destination Airport", property: "dest" },
+];
+
+// Convert routes into "human readable format"
+const rows = routes.map((route) => {
+  return {
+    airline: getAirlineById(route.airline, airlines).name,
+    src: getAirportByCode(route.src, airports).name,
+    dest: getAirportByCode(route.dest, airports).name,
+  };
+});
 
 const App = () => {
   return (
@@ -9,7 +27,12 @@ const App = () => {
         <h1 className="title">Airline Routes</h1>
       </header>
       <section>
-        <Table className="routes-table" />
+        <Table
+          className="routes-table"
+          columns={columns}
+          rows={rows}
+          format=""
+        />
       </section>
     </div>
   );
